pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opinion-hold-backend"
version = "0.1.0"
description = "FastAPI backend that surfaces OPINION markets near expiry with edge price >= 95%"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "websockets>=12.0,<16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "frontend*"]
namespaces = false

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
